Legg behandlingFakta inn i behandlingcontext
