+++ src/frontend/Sider/Behandling/BehandlingContainer.tsx
@@ -7,6 +7,7 @@ import { useApp } from '../../context/AppContext';
 import { useRerunnableEffect } from '../../hooks/useRerunnableEffect';
 import DataViewer from '../../komponenter/DataViewer';
 import { Behandling } from '../../typer/behandling/behandling';
+import { BehandlingFakta } from '../../typer/behandling/behandlingFakta/behandlingFakta';
 import { Personopplysninger } from '../../typer/personopplysninger';
 import { byggTomRessurs, Ressurs } from '../../typer/ressurs';
 
@@ -15,6 +16,8 @@ const BehandlingContainer = () => {
     const [behandling, settBehandling] = useState<Ressurs<Behandling>>(byggTomRessurs());
     const [personopplysninger, settPersonopplysninger] =
         useState<Ressurs<Personopplysninger>>(byggTomRessurs());
+    const [behandlingFakta, settBehandlingFakta] =
+        useState<Ressurs<BehandlingFakta>>(byggTomRessurs());
 
     const behandlingId = useParams<{
         behandlingId: string;
@@ -36,13 +39,22 @@ const BehandlingContainer = () => {
         document.title = 'Behandling';
     }, []);
 
+    const hentBehandlingFaktaCallback = useCallback(() => {
+        request<BehandlingFakta, null>(`/api/sak/behandling/${behandlingId}/fakta`).then(
+            settBehandlingFakta
+        );
+    }, [request, behandlingId]);
+
+    useEffect(hentBehandlingFaktaCallback, [hentBehandlingFaktaCallback]);
+
     return (
-        <DataViewer response={{ behandling, personopplysninger }}>
-            {({ behandling, personopplysninger }) => (
+        <DataViewer response={{ behandling, personopplysninger, behandlingFakta }}>
+            {({ behandling, personopplysninger, behandlingFakta }) => (
                 <BehandlingInnhold
                     behandling={behandling}
                     hentBehandling={hentBehandling}
                     personopplysninger={personopplysninger}
+                    behandlingFakta={behandlingFakta}
                 />
             )}
         </DataViewer>
