+++ src/frontend/Sider/Behandling/BehandlingInnhold.tsx
@@ -10,6 +10,7 @@ import { VilkårProvider } from '../../context/VilkårContext';
 import { RerrunnableEffect } from '../../hooks/useRerunnableEffect';
 import PersonHeader from '../../komponenter/PersonHeader/PersonHeader';
 import { Behandling } from '../../typer/behandling/behandling';
+import { BehandlingFakta } from '../../typer/behandling/behandlingFakta/behandlingFakta';
 import { Personopplysninger } from '../../typer/personopplysninger';
 
 const BehandlingContainer = styled.div`
@@ -26,9 +27,14 @@ const BehandlingInnhold: React.FC<{
     behandling: Behandling;
     hentBehandling: RerrunnableEffect;
     personopplysninger: Personopplysninger;
-}> = ({ behandling, hentBehandling, personopplysninger }) => {
+    behandlingFakta: BehandlingFakta;
+}> = ({ behandling, hentBehandling, personopplysninger, behandlingFakta }) => {
     return (
-        <BehandlingProvider behandling={behandling} hentBehandling={hentBehandling}>
+        <BehandlingProvider
+            behandling={behandling}
+            hentBehandling={hentBehandling}
+            behandlingFakta={behandlingFakta}
+        >
             <PersonopplysningerProvider personopplysninger={personopplysninger}>
                 <PersonHeader fagsakPersonId={behandling.fagsakPersonId} />
                 <BehandlingContainer>
