+++ src/frontend/typer/behandling/behandlingFakta/behandlingFakta.ts
@@ -4,6 +4,7 @@ import { FaktaDokumentasjon } from './faktaDokumentasjon';
 import { FaktaHovedytelse } from './faktaHovedytelse';
 
 export interface BehandlingFakta {
+    søknadMottattTidspunkt?: string;
     hovedytelse: FaktaHovedytelse;
     aktivitet: FaktaAktivtet;
     barn: FaktaBarn[];
